Make removing files parentReference-aware and delete instead of trash
Files can have multiple parentReferences, meaning the same fileId and
name can show up as a childReference in more than one folder. However,
outright deleting the fileId would remove it from both folders which
might result in an unintended loss of data.

Therefore, we first check if the file has more than one parent - when
this is the case we find the parentId of the current folder and remove
that parentId from the fileId.

If the fileId has exactly one parentId, we delete the entire fileId
rather than removing the final parentId - which would result in the
fileId becoming an invisible "ghost" in your OneDrive.

See also: https://phabricator.kde.org/T3443